Move API batch details definition to common library to ensure consistency between client and service

+++ clients/src/main/java/com/forsvarir/file/utils/fileclient/services/FileDetailsService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class FileDetailsService {
-    public void createFile(int batchId, FileDetails fileDetails) {
+    public void createFile(long batchId, FileDetails fileDetails) {
         System.out.println(fileDetails.getPath() + '/' + fileDetails.getName());
     }
 }
